# Buffer write improved