We've noticed that writing a lot in the socket for buffer operations took a lot time to process and we've developed a "flexbuffer" module which helps to accomplish such task.

We implemented in the node_redis client for buffer operations and benchmarks look quite good.

Below my results:

SET small buf: 5201 -> 9009 ops/sec (∆ +3808 )
SET small buf: 8054 -> 18281 ops/sec (∆ +10227 )
SET small buf: 8873 -> 19193 ops/sec (∆ +10320 )
SET small buf: 8539 -> 16488 ops/sec (∆ +7949 )


SET large buf: 6337 -> 7443 ops/sec (∆ +1106 )
SET large buf: 8163 -> 14174 ops/sec (∆ +6011 )
SET large buf: 10493 -> 18744 ops/sec (∆ +8251 )
SET large buf: 11487 -> 11560 ops/sec (∆ +73 )

#### Ready to merge checklist
- [x] passes tests in test.js
- [ ] merges cleanly
- [ ] coding style (4-space indents, looks similar to other code)