@bcoe 2015-08-18
@marcosnils this ticket is pretty ancient, so I'm closing it as part of the cleanup we're currently taking on. Please fee free to reopen if you'd like to rebase and work on getting this merged.